Upgrade USPS_Online::RateRequest to use Moose.
Make minimal changes necessary to get the few basic tests in 30_USPS_Online.t to execute. Uses BUILD() to set the default from_country to 'US'.

Add some tests to exercise basic functionality. Convert tests to use one block for each test or group of related tests.

Change "Flat Rate" to "Flat-Rate" for compatibility with USPS.

One of my last changes was a mistake:
> One, for example, is that virtual base classes do not have virtual
> objects that their concrete classes will have. That is, ::RateRequest does
> not have ::Shipment, because UPS_Online::  RateRequest will have its own
> UPS_Online::Shipment (inherited from ::RateRequest and ::Shipment,
> respectively).

The unintended side effect of that is you are now required to manually forward (i.e. set 'handles') for the methods that are normally handled by the parent (or grandparent) class through a virtual sub-object. It may not be the most obvious or straightforward design.