Reduce Log4perl indirection by yet another level.
The use of caller(2) was inconsistent, as warned in perldoc -f caller. Changed to a yet-shorter, thinner, and faster calling structure (wrapper) for Log4perl, now allowing the use of caller(1).

Tune configuration file. Document config file and logging functions.

Fix the log_level() sub so it can check the true levels.