Fix bug in shipping to United Kingdom.
Fix a bug where the CountryCode was always being set to 'UK', when UPS needs
their 'GB' value. The cause of the issue is that the Config::IniFiles module
started returning scalar string values when there is only 1 value in the
configuration, instead of an arrayref of one value like before. We work
around the issue by manually checking for and adding an arrayref if
necessary.