Add UPS Negotiated Rates support thanks to Paul Trader.
This feature allows UPS customers that have negotiated rates to get those
back from the UPS_Online module. It requires an abbreviated state code, but
it's probably good to add it to all requests anyway, so that is done (by
re-using the same methods that were created for UPS_Offline).

Patch applied as-is. Minor fixes and changes coming in future commits.